requirements.txt
pypresence
python-dotenv
requests